Grid search optimization for PCM runout distance

Computes performance measures for runout paths simuluated using the random walk and PCM model components of the GPP tool in SAGA-GIS.

pcmGridsearch(
  dem,
  slide_plys,
  slide_src,
  slide_id = NULL,
  rw_slp,
  rw_ex,
  rw_per,
  pcm_mu_v,
  pcm_md_v,
  gpp_iter = 1000,
  buffer_ext = 500,
  buffer_source = NULL,
  predict_threshold = 0.5,
  save_res = FALSE,
  plot_eval = FALSE
)

Arguments

dem
A DEM as a RasterLayer object

slide_plys
Runout tracks as a SpatialPolygonsDataFrame

slide_src
Source points as a SpatialPointsDataFrame or source areas as a SpatialPolygonsDataFrame

slide_id
Selects a single runout polygon from slide_plys by row

rw_slp
Random walk slope threshold - below lateral spreading is modelled

rw_ex
Random walk exponent controlling lateral spread

rw_per
Random walk persistence factor to weight flow direction consistency

pcm_mu_v
A vector of PCM model sliding friction coefficients

pcm_md_v
A vector of PCM model mass-to-drag ratios (m)

gpp_iter
Number of model iterations

buffer_ext
(Optional) Defines buffer distance (in meters) around runout polygon to crop source DEM. This helps to reduce computational time when working with large regions.

buffer_source
(Optional) Can define a buffer distance (in meters) to extend source point to a source area

predict_threshold
A cutoff value to define what quantile of simulated runout frequencies is the predicted runout.

save_res
(Logical) if TRUE, will save results in current working directory

plot_eval
(Logical) ff TRUE, will plot simulated runout and runout polygon

Value

A list with performances for each parameter combinations across grid search space. The performances measures include relative error, relative difference, error and AUROC.