PCM runout distance performance

Computes the error for runout distances simuluated using the random walk and PCM model components of the GPP tool in SAGA-GIS.

pcmPerformance(
  dem,
  slide_plys,
  slide_src,
  slide_id = 1,
  rw_slp = 33,
  rw_ex = 3,
  rw_per = 2,
  pcm_mu = 0.3,
  pcm_md = 75,
  buffer_ext = 500,
  buffer_source = 50,
  gpp_iter = 1000,
  predict_threshold = 0.5,
  plot_eval = FALSE,
  return_features = FALSE
)

Arguments

dem
A DEM as a RasterLayer object

slide_plys
Runout tracks as a SpatialPolygonsDataFrame

slide_src
Source points as a SpatialPointsDataFrame or source areas as a SpatialPolygonsDataFrame

slide_id
Selects a single runout polygon from slide_plys by row

rw_slp
Random walk slope threshold - below lateral spreading is modelled

rw_ex
Random walk exponent controlling lateral spread

rw_per
Random walk persistence factor to weight flow direction consistency

pcm_mu
PCM model sliding friction coefficient

pcm_md
PCM model mass-to-drag ratio (m)

buffer_ext
(Optional) Defines buffer distance (in meters) around runout polygon to crop source DEM. This helps to reduce computational time

buffer_source
(Optional) Can define a buffer distance (in meters) to extend source point to a source area

gpp_iter
Number of model iterations

predict_threshold
A cutoff value to define what quantile of simulated runout frequencies is the predicted runout.

plot_eval
(Logical) if TRUE, will plot simulated runout and runout polygon

return_features
(Logical) if TRUE, returned list will include GPP input and output data, in addition to a list of error measures.

Value

A list of runout distance performance measures.